remove_huita_add_size_t_easy_peasy_fix

+++ ft_striteri.c
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	unsigned int i;
+	size_t i;
 
 	if (s == NULL || f == NULL)
 		return ;
